fix: separate layout and container
